# Child spawning optimizations