--- a/benches/benches/bevy_hierarchy/main.rs
+++ b/benches/benches/bevy_hierarchy/main.rs
@@ -1,0 +1,5 @@
+use criterion::criterion_main;
+
+mod spawn;
+
+criterion_main!(spawn::benches);