# Objective

- Fixes some of #17301 

## Solution

- Avoid re-parenting and related checks when spawning new entities, which cannot exist in the hierarchy yet.

## Testing

- Added a `spawn_children` benchmark

---

## Showcase

Fixed exponential time complexity when spawning children on an entity that already has existing children. This case can now complete in linear time.

```
spawn_children/10_entities
                        time:   [1.4008 µs 1.4195 µs 1.4392 µs]
                        change: [-7.2431% -5.7892% -4.2650%] (p = 0.00 < 0.05)
spawn_children/100_entities
                        time:   [10.150 µs 10.302 µs 10.461 µs]
                        change: [-22.516% -21.350% -20.218%] (p = 0.00 < 0.05)
spawn_children/1000_entities
                        time:   [88.756 µs 89.768 µs 90.867 µs]
                        change: [-75.726% -75.512% -75.309%] (p = 0.00 < 0.05)
spawn_children/10000_entities
                        time:   [912.39 µs 924.42 µs 936.40 µs]
                        change: [-96.697% -96.659% -96.618%] (p = 0.00 < 0.05)
spawn_children/100000_entities
                        time:   [9.3389 ms 9.3936 ms 9.4529 ms]
                        change: [-99.655% -99.653% -99.651%] (p = 0.00 < 0.05)
spawn_children/1000000_entities
                        time:   [111.12 ms 111.84 ms 112.58 ms]
```

Note the 1 million case takes 111ms on this branch. On `main`, I had to kill the test because it was estimated to take 8 hours.