@aevyrie 2025-01-13
I have a few more changes locally I need to push. I noticed that some of the child methods send a hierarchy event, but some do not. I can't tell if this is an error or not. The docs don't really spell out any intent.