@aevyrie 2025-01-15
Can anyone with more familiarity please comment on the `HierarchyEvent`s? It seems like it is outright missing in a few places but I'm not sure if this is a bug or not.